+++ CHANGELOG.md
@@ -16,5 +16,6 @@
 - User model and tests
 - Add dry config
 - Get document (controller, contract, use_case and presenter)
+- Role model and tests
 
 ## Version [0.1.0.beta] - 2025-01-21